Added some more features. Adjusted how output is done.
Added a verbosity feature with the add_msg function. The class now defaults to the bare minimum output. Added commandline flags that will allow you to add and remove subreddits from the ban list as well as turn verbosity on and off. Also added more error handling to the process and subreddit checks where needed.